On branch addedMoreTestCases  Your branch is up-to-date with 'instasize-automation/addedMoreTestCases'.
 Changes to be committed:
	deleted:    .idea/workspace.xml
	new file:   100755
	modified:   Asserts.pyc
	modified:   DriverBuilder7zero.pyc
	modified:   InstasizePages.pyc
	renamed:    apps/OsakaFilterExport.py -> OsakaFilterExport.py
	new file:   PaidFiltersPage.pyc
	modified:   TryExcepts.pyc
	deleted:    apps/InstaSize_R9_Test_3_7_6_107_collage_crash_fix_3.apk
